requirements.txt
numpy>=1.26,<2.0
python-dotenv>=1.0
click>=8.1
pandas>=2.2
scikit-learn>=1.5
PyMySQL>=1.1
cryptography>=42.0
torch>=2.2,<2.5
mysql-connector-python>=9.0

# --- Optional: web scraping utilities (comment out if unused) ---
requests==2.32.2
beautifulsoup4==4.12.3
selenium==4.21.0
pytest>=8.0
sqlalchemy>=2.0
matplotlib>=3.8
streamlit==1.45.1
optuna>=3.6
